djangochatapp/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
pytz==2023.3
redis==5.0.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
zope.interface==6.0
